wip(editor): move mapleaflet-files to dedicated folder
